Move KnownIntrinsics to guest staging
Relocate the guest-safe intrinsic declarations without changing their graph-builder implementations. Keep readHub in a temporary core-owned declaration because DynamicHub is not guest-visible yet; GR-78110 tracks completing that move.
