+++ substratevm/src/com.oracle.svm.guest.staging/src/com/oracle/svm/guest/staging/core/graal/KnownIntrinsics.java
@@ -22,13 +22,12 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package com.oracle.svm.core.snippets;
+package com.oracle.svm.guest.staging.core.graal;
 
 import org.graalvm.nativeimage.c.function.CodePointer;
 import org.graalvm.word.Pointer;
 
 import com.oracle.svm.shared.NeverInline;
-import com.oracle.svm.core.hub.DynamicHub;
 
 /**
  * Functions that are implemented as compiler intrinsics. For implementation see
@@ -42,15 +41,10 @@ public class KnownIntrinsics {
     public static native Pointer heapBase();
 
     /**
-     * Returns the value of the code base, which is the address which
-     * {@linkplain com.oracle.svm.core.meta.MethodOffset method offsets} are relative to.
+     * Returns the value of the code base, which is the address which {@code MethodOffset} method
+     * offsets are relative to.
      */
     public static native Pointer codeBase();
-
-    /**
-     * Returns the hub of the given object.
-     */
-    public static native DynamicHub readHub(Object obj);
 
     /**
      * Narrow down the range of values to exclude 0 as the possible pointer value.
